samples/pyproject.toml
[tool.poetry]
name = "dftidy-demo"
version = "0.1.0"
description = "Demo project loading a CSV and running dftidy pipeline"
authors = ["Andy Rappaport andy@rappaport.name"]
readme = "README.md"
packages = [{ include = "dftidy_sample" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pandas = ">=2.0"
dftidy = { git = "https://github.com/arappaport/dftidy.git", branch = "main" }
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
ruff = ">=0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
